h2: make guest struct alloc'ed by lib [1/2]
If the struct is allocated outside of the library it isn't clear the
user should call h2_guest_free(), which is responsible to free memory of
pointers inside the struct. Therefore the user will be leaking memory
without noticing. By making the struct alloc'ed by the library it is
hopefully clearer that the user should call the correspondent free
function therefore free'ing all memory.

This patch changes the user facing interface. Next patch will take care
of changing the hypervisor specific code.