Makefile: fix paths to build from compiled xen src
Unfortunately libxc's xc_dom.h is not installed with the rest of xen
headers. This header is essential to create domains using libxc and
therefore we need to include it directly from the xen sources.

It is also not possible to build purely against xen sources due to a
couple of things: (1) include paths on some xen headers assume the
installation paths as opposed to the paths under the xen sources; (2)
some headers are moved/linked during the build process making it
possible to include them from chaos/libh2 (e.g. libelf.h). Therefore it
is required to build xen before building chaos.

With this commit is possible to build xen entirely from the compiled
sources without requiring it to be installed. You might need to
configure `XEN_PREFIX` under .config, if you don't use the default.