chaos: first basic implementation of the chaos
This is the basic client of libh2 that will provide basic, xl like,
functionality. The first implementation supports create and destroy VMs,
for xen, and using xc and xenstore. Create uses json files for
configuration, and destroy uses domain ids.

The client will evolve with the library, as new functionality arrives
there.