Allow memory pre-initialization
This allows to create a domain and reserve its memory before loading the
kernel, which allows doing this ealier (potentially much earlier) than
the actual kernel loading and booting of the domain (what you would
generally consider "creating/starting the domain").

Since the memory registration/initialization routines are specific to
the architecture, the address_size variable in struct guest needs to be
set (to either 32 or 64) before doing the memory initialization.