xen: add default console to guest
On xen the guest has a default console whose details are added to the
start_info page. Because of this and the way this is implemented by
libxc, the console device can only be created after the domain is
created which causes problems for the split domain instantiation.

Since this console is special anyway make it part of the guest
definition instead of a normal device. This way devices can be created
before the fastboot part of domain creation, an only console needs to be
created in the end.

This commit removes the console device type. This can latter be added
when support for multiple consoles is added.

This fixes #17.