chaos: command line option to skip contacting shell daemon
Trying to contact the shell daemon incurs a small overhead. If the user
knows that contacting it is useless (because it is not running, or
because the provided shells don't fit the requirements), or if they want
to skip fastbooting for whatever other reason, the --skip-daemon (-s)
command line option is now available to the "create" subcommand of the
chaos command line interface.