xen: fix domain memory size querying
Most of the times (always?) you want to query the memory allocated to
the domain by the user and not what the domain is really using at a
specific point in time, so max_pages should be used here.

Later we can add the currently allocated memory to the struct, but is
not important for now.