shell_daemon: don't close connection after every creation request
Instead, wait for a read error, which occurs when the sender closes its
connection.

Yes, this has the potential for a DoS in its current iteration, if chaos
never closes its connection to the daemon, then others cannot connect.
But since only root can talk to the daemon, this is not a huge deal.