Correct license headers
This commit removes the additional clause
"THIS HEADER MAY NOT BE EXTRACTED OR MODIFIED IN ANY WAY"
in the license headers that got introduced by mistake.

On the one hand, this additional clause is redundant because
the referring license already states that it must remain and
the copyright notice must be kept.
On the other hand, the clause freezes the file header
and prohibits future contributors from extending the
copyright notice for their contributions. This additional
clause is not part of the official BSD license where the
license in the affected files is derived from.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>